Demographic data on study participants.

Demographic data on study participants, such as population, age, and education.

mysore2015dem

Format

A data frame with 102 rows and 7 variables:

ID

double. Participant ID.

Population

character. Villager or Jenu Kuruba.

Age

double. Age in years. This variable is omitted to maintain anonymity, but will be provided to researchers who agree to the terms of our IRB.

months.preg

double. Months pregnant.

trimester

double. Trimester

Education

double. Number of years of education.

birthno

double. Number of births.

Source

placekmysore2015publicdata