Explanatory variables for the study

A dataset containing the explanatory variables, such as food insecurity, perceived susceptibility to disease, and so forth. Food insecurity items are from the 6-item short-form household-level food insecurity measure, which assesses one’s access to sufficient foods (Blumberg, Bialostosky, Hamilton, & Briefel, 1999).

mysore2015exp

Format

A data frame with 102 rows and 28 variables:

ID

Subject ID

nausea

Current nausea (0 = no, 1 = yes).

vomit

Current vomiting (0 = no, 1 = yes).

HWWStoilet

do you hand wash with soap after using toilet? (0 = no, 1 = yes).

HWWScleanchild

do you hand wash with soap after cleaning child? (0 = no, 1 = yes).

HWWSfeedchild

do you hand wash with soap before feeding child? (0 = no, 1 = yes).

HWWSfood

do you hand wash with soap before handling food? (0 = no, 1 = yes).

immune1

In general, I am very susceptible to colds, flu and other infectious diseases (0 = no, 1 = yes).

immune2

My immune system protects me from most illnesses that other people get (0 = no, 1 = yes).

immune3

I have a history of susceptibility to infectious disease (0 = no, 1 = yes).

toilet

type of toilet: none (0) or squat (1)

waterclean

perceived cleanliness of drinking water: poor (0), good (1).

dietadvice

who do you ask for advice regarding diet in pregnancy

dietadvice2

who do you ask for advice regarding diet in pregnancy

dietadvice3

who do you ask for advice regarding diet in pregnancy

dietadvice4

who do you ask for advice regarding diet in pregnancy

pressure

does anyone pressure to follow certain guidelines in pregnancy? (first person listed)

pressure2

2nd person listed that gives pressure to follow pregnancy guidelines

pressure3

3rd person listed that gives pressure to follow pregnancy guiedlines

totalhousehold

total number of individuals living in house

friend.sis.pregnantYN

Are any of your friends or sisters pregnant?

fs1

food insecurity item 1

fs2

food insecurity item 2

fs3

food insecurity item 3

fs4

food insecurity item 4

fs5

food insecurity item 5

fs6

food insecurity item 6

fstotal

total food insecurity score

Source

placekmysore2015publicdata