Free-listed food avoidances and aversions

A dataset containing free-listed food avoidances and aversions, the symptoms experienced if consuming aversive foods, and the perceived consequences of consuming foods that are taboo.

mysore2015food

Format

A data frame with 102 rows and 69 variables:

ID

participant ID

aversion

first item listed as pregnancy aversion

symptom

symptom experienced from first mentioned aversion

aversion2

second item listed as pregnancy aversion

symptom2

symptom experienced from second mentioned aversion

aversion3

third item listed as pregnancy aversion

symptom3

symptom experienced from third mentioned aversion

aversion4

fourth item listed as pregnancy aversion

symptom4

symptom experienced from fourth mentioned aversion

aversion5

fifth item listed as pregnancy aversion

symptom5

symptom experienced from fifth mentioned aversion

avoidance1

first item listed as pregnancy avoidance

learn1

how individual learned about food avoidance

consequence1

consequence of consuming forbidden item

avoidance2

second item listed as pregnancy avoidance

learn2

how individual learned about food avoidance

consequence2

consequence of consuming forbidden item

avoidance3

third item listed as pregnancy avoidance

learn3

how individual learned about food avoidance

consequence3

consequence of consuming forbidden item

avoidance4

fourth item listed as pregnancy avoidance

learn4

how individual learned about food avoidance

consequence4

consequence of consuming forbidden item

avoidance5

fifth item listed as pregnancy avoidance

learn5

how individual learned about food avoidance

consequence5

consequence of consuming forbidden item

avoidance6

sixth item listed as pregnancy avoidance

learn6

how individual learned about food avoidance

consequence6

consequence of consuming forbidden item

avoidance7

seventh item listed as pregnancy avoidance

learn7

how individual learned about food avoidance

consequence7

consequence of consuming forbidden item

fruit_avoid

logical. Does the participant avoid at least one fruit?

meat_avoid

logical. Does the participant avoid at least one meat?

vegetable_avoid

logical. Does the participant avoid at least one vegetable?

nut.seed.legume_avoid

logical. Does the participant avoid at least one nut/seed/legume?

grains_avoid

logical. Does the participant avoid at least one grain?

ess_avoid

logical. Does the participant avoid at least one ethnic strong and spicy food?

dairy_avoid

logical. Does the participant avoid at least one dairy food?

fruit_averse

logical. Is the participant averse to at least one fruit?

meat_averse

logical. Is the participant averse to at least one meat?

vegetable_averse

logical. Is the participant averse to at least one vegetable?

nut.seed.legume_averse

logical. Is the participant averse to at least one nut/seed/legume?

ess_averse

logical. Is the participant averse to at least one ethnic strong and spicy food?

grains_averse

logical. Is the participant averse to at least one grain?

sweets_averse

logical. Is the participant averse to at least one sweet?

Source

https://doi.org/10.5281/zenodo.836843